app/requirements.txt
djangorestframework==3.14.0
Django==5.0.2
PyJWT==2.8.0
django-cors-headers==4.3.1